separation of HTTP specific implementations and general API. refactored OAuth. Some package changes

+++ core/src/main/java/org/codegist/crest/annotate/ConnectionTimeout.java
@@ -36,6 +36,6 @@ import java.lang.annotation.Target;
 @Target({ElementType.TYPE,ElementType.METHOD})
 public @interface ConnectionTimeout {
 
-    long value();
+    int value();
 
 }
